requirements.txt
django
markdown
gunicorn
leapcell
python-dotenv